requirements.txt
Django==4.0.6
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
whitenoise==6.2.0
psycopg2-binary==2.9.3

djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.0

japanize_matplotlib=1.1.3
matplotlib=3.5.2
numpy=1.23.1
pyairtable=1.2.0
pycryptodomex=3.15.0
python-dateutil=2.8.2
scipy=1.9.0
